refactor: распределяет код по новым модулям
